ANOTHER HUGE UPDATE IS ON THE WAY 😉
-file hierarchy edited.
--After I get a successful build, I thought like, it was easier than I thought, then look at the time and see it passed ~4 hours 🤠
-GetFileContents function moved to input.
-input devices changed to structs
-opengl objects (vbo, vao, ibo, shader) changed to structs and function definitions moved to header files.
-I want to be consistent and if the object is just a tool like a vector, array object etc. and have functions which effecting only itself, it will be a struct.
-vertex, face, edge and meshdata structs are added to toolkit.
-MeshBuilder class added to toolkit and it will be all static, maybe I can move it somewhere else
-mesh data I want to make uses the render dynamic representation in this article:
https://en.wikipedia.org/wiki/Polygon_mesh
-mesh will support only triangles because faces in this system takes only 3 edge and vertex pointers.
-responsibilty of setting up and sending all vertex and index data to gpu is entirely belongs to Mesh component. but getting a vertex cloud and index array and convertin it to mesh data is belongs to MeshBuilder.

+++ GamEncin/src/Encin/Items/Application.cpp
@@ -1,10 +1,7 @@
-#include "Encin.h"
-#include <stdio.h>
-#include <time.h>
+#include "Encin/Encin.h"
 
 namespace GamEncin
 {
-
 #pragma region Variable Definitions
 
     Scene* Application::currentScene = nullptr;
@@ -78,7 +75,7 @@ namespace GamEncin
         if(printFPS)
         {
             char buff[100];
-            snprintf(buff, sizeof(buff), "%s | FPS: %f | Delta Time: %f", programName.c_str(), 1 / deltaTime, deltaTime);
+            snprintf(buff, sizeof(buff), "%s | FPS: %d | Delta Time: %f", programName.c_str(), (int) (1 / deltaTime), deltaTime);
             glfwSetWindowTitle(currentScene->renderer->window, buff);
         }
     }
